@@ -2,8 +2,8 @@
 	"name": "echomods",
 	"private": true,
 	"version": "0.1.7",
-	"main": "main.js",
-	"type": "commonjs",
+	"main": "main.mjs",
+	"type": "module",
 	"author": {
 		"name": "Andrei Hudalla"
 	},
@@ -88,7 +88,7 @@
 		"@electron-forge/plugin-auto-unpack-natives": "^6.4.2",
 		"@iconify/vue": "^4.1.1",
 		"@vitejs/plugin-vue": "^4.1.0",
-		"electron": "^24.8.5",
+		"electron": "^28.0.0",
 		"electron-builder": "^24.9.1",
 		"electron-wix-msi": "^5.0.0",
 		"esm": "^3.2.25",
